bug: fixes for ForceHessian and printouts
The ForceHessian displacements should be largest
for the lightest atoms and smallest for the heaviest
atoms. I had reversed it.

Added a diff on Array1D to return the difference.

Fixed import of the correct array classes. Now
they have all the functionality because of the
correct requires.

Updated the print-out of the LBFGS algorithm.

Added the lattice optimization algorithm.
It currently has not been tested.

Signed-off-by: Nick Papior <nickpapior@gmail.com>