enh: changed array to arbitrary length array
Instead of having two classes for each array-dimension type we
now have a single array which makes everything MUCH easier.

Indeed the implementations are also much simpler to work with.
To utilize the new method one may simply do:

  num = require "flos.num"

and do
  a = num.Array( 2, 4 )
or
  a = num.Array( 2 )

Nearly all mathematical operations are present with the limitations of
dot products of more than 2 dimension arrays.

Signed-off-by: Nick Papior <nickpapior@gmail.com>