enh: added the CG method for relaxation
Currently this is a minimal CG method and one may choose any
optimizer (other than CG) for performing the line-minimization.
The line-minimizer then optimizes the projected gradient onto the direction
which is equivalent to minimizing along the direction vector.

Sadly the CG method, as in regular SIESTA shows a fast initial convergence
and then a _very_ slow final descent. Perhaps an initial CG relaxation and
then subsequent a LBFGS relaxation for fast relaxation?

Signed-off-by: Nick Papior <nickpapior@gmail.com>